Port to new approach
This removes some code duplication, uses a bit more of an S3 style to do the
conversions, and separates out the summary logic from the construction logic.

I've made a start on the tests.  Plotting needs a lot of work but I think
that @thibautjombart has more ideas what needs doing here.

README still needs adding as does a vignette.