adding new incidence_fit_list class
This is to deal with the new `separate_split`
option in `fit_optim_split()`. At the moment,
the result of this is to return a list with
groups at the top level and the incidence_fit
object nested far below like so:

    x$<group>$fit$<before/after>

This class implements an attribute that stores
a list of the paths so that they can be used for
subsetting like so:

attr(x, "locations")
 males, fit, before
 males, fit, after
 females, fit, before
 females, fit, after

This takes advantage of the fact that you can
traverse nested lists with a vector specifying
where you belong at each given level.